Moved math.hpp to detail
