+++ include/boost/simd/arch/common/generic/function/sqrt.hpp
@@ -19,7 +19,7 @@
 #include <boost/simd/function/is_gez.hpp>
 #include <boost/simd/function/multiplies.hpp>
 #include <boost/simd/function/rsqrt.hpp>
-#include <boost/simd/math.hpp>
+#include <boost/simd/detail/math.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/assert.hpp>
 #include <boost/config.hpp>
