+++ include/boost/simd/arch/common/scalar/function/abs.hpp
@@ -13,7 +13,7 @@
 #define BOOST_SIMD_ARCH_COMMON_SCALAR_FUNCTION_ABS_HPP_INCLUDED
 #include <boost/simd/function/std.hpp>
 
-#include <boost/simd/math.hpp>
+#include <boost/simd/detail/math.hpp>
 #include <boost/simd/function/std.hpp>
 #include <boost/dispatch/meta/as_unsigned.hpp>
 #include <boost/dispatch/function/overload.hpp>
