+++ include/boost/simd/arch/common/scalar/function/copysign.hpp
@@ -17,7 +17,7 @@
 #include <boost/simd/function/scalar/bitofsign.hpp>
 #include <boost/simd/function/scalar/signnz.hpp>
 #include <boost/simd/constant/signmask.hpp>
-#include <boost/simd/math.hpp>
+#include <boost/simd/detail/math.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/config.hpp>
