+++ include/boost/simd/arch/common/scalar/function/round.hpp
@@ -21,7 +21,7 @@
 #include <boost/simd/function/scalar/is_ltz.hpp>
 #include <boost/simd/function/scalar/seldec.hpp>
 #include <boost/simd/function/scalar/tenpower.hpp>
-#include <boost/simd/math.hpp>
+#include <boost/simd/detail/math.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/config.hpp>
 #include <cmath>
