+++ include/boost/simd/arch/common/scalar/function/sqrt.hpp
@@ -16,7 +16,7 @@
 #include <boost/simd/function/std.hpp>
 
 #include <boost/simd/function/scalar/is_gez.hpp>
-#include <boost/simd/math.hpp>
+#include <boost/simd/detail/math.hpp>
 #include <boost/simd/function/std.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/assert.hpp>
