+++ include/boost/simd/detail/math.hpp
